Modify playGame() to loop and announce winner
The for loop and the announce winner algorithm is now placed inside the
function playGame() itself instead of in the global scope.
The user now has to call playGame() in the console to play the game
since it was not invoked or called in the global scope to initiate the
prompt box.